ospfd: Fix crash on shutdown in maxage_lsa table
The prefix structure used for lookup on arm platforms
was not being zero'ed out before adding data.  Therefore
when we needed to lookup the same lsa we would get
a different hash bucket.  This caused upon shutdown
the lsa pointer to be free'd too many times since
we had multiple pointers to it.

Ticket: CM-17706
Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>
Signed-off-by: Don Slice <dslice@cumulusnetworks.com>
Signed-off-by: Daniel Walton <dwalton@cumulusnetworks.com>