zebra: install static routes in a vrf when next-hop interface comes up
Problem reported with cross-vrf static routes that the routes weren't
installed when the target interface is bounced.  Determined that we did
not initiate re-install of the statics in that particular case, so added
it.  Manual testing attached to the defect.  Test case previously failing
now passes.  Reset of VRF automated tests and bgp-min will be completed
before pushing.

Ticket: CM-22828
Signed-off-by: Don Slice <dslice@cumulusnetworks.com>