zebra: vxlan intf update add info to debug log
Added interface state and previous assigned vtep-ip
as part of the debug to rc cm-22929 where clag peerlink
flap, individual vtep is is not withdrawn.

One reason is when interface update came with chagflag
IP address change, ifp might be in down state and fails
to send update to bgp to withdraw old vtep-ip.
The issue is intermittant and this log can help
rc the issue when the issue is hit.

Signed-off-by: Chirag Shah <chirag@cumulusnetworks.com>