bgpd: Refine check for which routes can be leaked into VPN
A non-imported route or a non-VPN (e.g., EVPN) imported route is a
candidate to be exported into the VPN routing table for leaking to
other BGP instances or advertisement into L3VPN. Implement a function
to check this and use in the appropriate places.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>
Reviewed-by:   Anuradha Karuppiah <anuradhak@cumulusnetworks.com>
Reviewed-by:   Donald Sharp <sharpd@cumulusnetworks.com>

Ticket: CM-23050
Reviewed By: CCR-8225
Testing Done:  Manual testing, evpn-smoke; evpn-smoke to be rerun.