zebra: Remove the re->prev pointer
The struct route_entry data structure has prev and next
for a custom linked list of data.  The prev data pointer is
useful if we plan to do a lot of insertion/deletions of route
entries, but the life cycle of route entries is long, adding
a bit extra work on route removal is ok in this case.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>