[CUMULUS] bgpd: Fix route change handling for EVPN-based routes
IPv4 and IPv6 routes based on EVPN require some special handling for
their associated next hop and its MAC. In the current implementation,
a change requires a delete followed by add to achieve the required
cleanup. This was not done in one situation, resulting in stale EVPN
next hops with associated stale (Router) MAC which were still being
incorrectly used even when the next hop's (Router) MAC had changed.
That is addressed with this commit.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>
Acked-by:      Anuradha Karuppiah <anuradhak@cumulusnetworks.com>

Note:
This is a sub-optimal solution, but that is already the case with the
implementation. This scenario will be optimized in 4.0, and then
considered for a future 3.7.x.

Ticket: CM-23892
Reviewed By:
Testing Done:
1. Manual testing
2. Verification in replication testbed by John Frazier.