lib: Add tail pointer to nexthop_group struct
Add a pointer to the last nexthop in the nexthop_group struct.
As long as the nexthop_group was created using the proper
_nexthop_group_add_sorted() call, the struct should maintain order
and keep the tail pointer updated. This helps so that when we get
a nexthop to add to the end, we can quickly access the end with
iterating over the nexthop linked list and getting a performance
hit.

Signed-off-by: Stephen Worley <sworley@cumulusnetworks.com>