pimd: Send correct number of BULK items down
The code was first checking to see if we should
buffer flush.  If the operation ended up being
a bulk add then it was incrementing the bulk item
counter.  Immediately after this we were deciding
if the number of entries was going to be too big
for our stream, and send it down.  If the new
data was going to be too big we would send immediately.

The problem here is that we were incrementing the count
but sending count - 1 down due to the stream check.

Move the stream check to *Before* the increment and
send it then.

Ticket: CM-27055
Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>
Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>