zebra: find/create an NHE via a single nexthop
Add an API to allow someone to find a nhg_hash_entry via
a single nexthop. This will also account fot hte case of
it being recursively resolved.

The AFI of the route using it MUST be passed if its a blackhole
nexthop.

This patch also uses this new API to deduplicate a lot of the
code in the depends helper functions as well as the lookup
via a single nexthop API.

Signed-off-by: Stephen Worley <sworley@cumulusnetworks.com>