zebra: nhg_depend more walk/num/exist functions
Expand the zebra_nhg_depends_* APIs with some more functions
for walking nexthops with a specified flag.

Expand the zebra_nhg_depends_* APIs with functions for counting
the number of depend nodes and nexthops in the whole tree starting
from specified root. With just counting nodes, we do not include self
in the count. With counting the number of lib/nexthops, we include
self. Otherwise, a single nexthop would be considered size 0 which
is wrong.

Expand the zebra_nhg_depends_* APIs to include with functions for
determining if a specified nexthop exists in the tree (match or ignore
labels options avaliable).

Signed-off-by: Stephen Worley <sworley@cumulusnetworks.com>