zebra: zebra_nhg API convert NHE to lib/nexthop_group
Add an API to `zebra/zebra_nhg.h` for converting the full
NHE depends tree into a traditional `lib/nexthop_group` struct
with the nexthop group linked list that other daemons and the
dataplane uses.

They don't care about the tree resolution so we need a way to convert
back to the lib nexthop_group for them to use.

Signed-off-by: Stephen Worley <sworley@cumulusnetworks.com>