zebra: add for_each macros for top level NHEs/nexthops
Add some for_each macros for iterating over the top level
nexthops in an NHE. This means we do not iterate over
the recursively resolved ones.

It is expected that you are passing in a group (not a singleton).

This is functionally simialr to this:

for (nexthop = *; nexthop; nexthop = nexthop->next) {}

Ex)
	1:
		-> 2:
			-> 3
		-> 4
		-> 5:
			-> 6
		-> 7

	for (2,4,5,7)

Signed-off-by: Stephen Worley <sworley@cumulusnetworks.com>