zebra: fix incorrect uplink oper count update
The operdown call back was executing multiple times on an ifdown causing
the uplink_oper_up to get decremented more than once, incorrectly.

This was resulting in an incorrect/permanent mh protodown rc after events
such as -
1. "systemctl networking restart".
2. ifdown/ifup of bond links

Fix is to maintain a uplink oper flag to ignore dups.

Ticket: CM-29512

Signed-off-by: Anuradha Karuppiah <anuradhak@cumulusnetworks.com>