pbrd: assume sequence installed/removed success until callback
Until we get a callback saying the rule was installed/removed
successfully, let's assume it was. Otherwise, we see weird
timing issues where we can send a delete then immediately get
an add that we never act on because we still think its installed
while waiting for the del response.

Signed-off-by: Stephen Worley <sworley@cumulusnetworks.com>