[CUMULUS] zebra: add --enable-nhg option for testing
Add --enable-nhg option for Cumulus testing before we
turn it on by default.

Internally, we disable NHG use by default except for the
multihoming use-case and sharpd. This option gives us
the ability to turn it on at runtime without compilation
so we can test it as well as let customers give it a
try.

Ticket: CM-32204

Signed-off-by: Stephen Worley <sworley@nvidia.com>