lib: remove unnecessary validation in plist code
libyang has severe performance issues with relative path interpretation.
We do relative path interpretation in order to manually enforce a 'when'
constraint that was also removed for performance reasons. However, this
vlaidation isn't necessary when using the CLI, as the northbound
callbacks to modify prefix lists will never be called unless the
prefix-list type (ipv4 or ipv6) is already known to be appropriate for
the prefix list we are modifying.

This won't necessarily be the case for API based northbound clients, and
the constraint needs to be enforced again in the future.

Signed-off-by: Quentin Young <qlyoung@nvidia.com>