lib: nb batching handle no form cmd
when a no form of the cli is part of the batched
commands, the consistent results (in case of prefix-list
config daemon crash) are observed.

solution:
In northbound batching framework, if no form of cmd
is present, perform following actions
- Flush any prior batched cmds
- Execute the no form of the cmd.

Ticket: CM-33109
Testing Done:

Executed 27k prefix-lists configuration with
single prefix-list with 27K rules,
then execute second file which contains no form
of the prefix-list then add modified version
of rules to the prefix-list.

Signed-off-by: Chirag Shah <chirag@nvidia.com>
Signed-off-by: Donald Sharp <sharpd@nvidia.com>